feat(optimize): remove image optimizations
- all image optimize modules removed
- responsive-image with responsive-image-static replaced
