+++ package.json
@@ -44,7 +44,6 @@
     "packtracker": "packtracker-upload --stats=reports/webpack/stats/modern.json --output-path=dist/_nuxt/"
   },
   "devDependencies": {
-    "@bazzite/nuxt-optimized-images": "0.4.0",
     "@commitlint/cli": "8.3.5",
     "@commitlint/config-conventional": "8.3.4",
     "@fullhuman/postcss-purgecss": "2.1.2",
@@ -73,9 +72,6 @@
     "eslint-plugin-security": "1.4.0",
     "express": "4.17.1",
     "husky": "3.1.0",
-    "imagemin-mozjpeg": "8.0.0",
-    "imagemin-webp": "5.1.0",
-    "imagemin-webpack-plugin": "2.4.2",
     "license-webpack-plugin": "2.1.4",
     "lint-staged": "10.0.5",
     "lost": "8.3.1",
@@ -89,9 +85,7 @@
     "postcss-object-fit-images": "1.1.2",
     "postcss-url": "9.0.0",
     "reqlim": "0.0.0",
-    "responsive-loader": "1.2.0",
     "rucksack-css": "1.0.2",
-    "sharp": "0.25.2",
     "storybook-addon-vue-info": "1.4.2",
     "stylelint": "13.3.3",
     "stylelint-config-css-modules": "2.2.0",
