+++ src/components/organisms/SinglePicture.vue
@@ -11,7 +11,7 @@
 
 <script>
 import LayoutDefaultContainer from '@/components/layouts/DefaultContainer';
-import AtomResponsiveImage from '@/components/atoms/ResponsiveImageStatic';
+import AtomResponsiveImage from '@/components/atoms/ResponsiveImage';
 
 export default {
   components: {
