+++ src/components/organisms/StagePicture.vue
@@ -55,7 +55,7 @@
 
 <script>
 import LayoutDefaultContainer from '@/components/layouts/DefaultContainer';
-import AtomResponsiveImage from '@/components/atoms/ResponsiveImageStatic';
+import AtomResponsiveImage from '@/components/atoms/ResponsiveImage';
 
 export default {
   components: {
